add websocket notifications

+++ src/transactions/transactions.service.ts
@@ -1,11 +1,23 @@
-import { BadRequestException, Injectable } from '@nestjs/common';
+import {
+    BadRequestException,
+    Injectable,
+    NotFoundException,
+} from '@nestjs/common';
 import { InjectModel } from '@nestjs/sequelize';
 import {
+    EVENT_TRANSACTION_CREATED,
+    EVENT_TRANSACTION_DELETED,
+    EVENT_TRANSACTION_UPDATED,
+    TRANSACTION_CREATED,
+    TRANSACTION_DELETED,
     TRANSACTION_NOT_DELETED,
+    TRANSACTION_NOT_FOUND,
     TRANSACTION_NOT_UPDATED,
+    TRANSACTION_UPDATED,
     TransactionTypes,
 } from 'src/common/constants';
 import { Transaction } from 'src/common/models/transactions.model';
+import { NotificationGateway } from 'src/notification/notification.gateway';
 import { RedisService } from 'src/redis/redis.service';
 import { UserService } from 'src/user/user.service';
 import { CreateTransactionDto } from './dto/create.dto';
@@ -17,6 +29,7 @@ export class TransactionsService {
         @InjectModel(Transaction) private transactionModel: typeof Transaction,
         private readonly userService: UserService,
         private readonly redisService: RedisService,
+        private readonly notificationGateway: NotificationGateway,
     ) {}
 
     async getAll() {
@@ -56,10 +69,21 @@ export class TransactionsService {
                 break;
         }
         await this.redisService.client.del('transactions');
-        return await this.transactionModel.create(dto);
+        const transaction = await this.transactionModel.create(dto);
+        this.notificationGateway.sendNotification(
+            dto.userId,
+            TRANSACTION_CREATED,
+            EVENT_TRANSACTION_CREATED,
+            transaction,
+        );
+        return transaction;
     }
 
     async updateById(id: string, dto: UpdateTransactionDto) {
+        const transaction = await this.getById(id);
+        if (!transaction) {
+            throw new NotFoundException(TRANSACTION_NOT_FOUND);
+        }
         const affectedCount = await this.transactionModel.update(dto, {
             where: {
                 id,
@@ -68,11 +92,22 @@ export class TransactionsService {
         if (!affectedCount) {
             throw new BadRequestException(TRANSACTION_NOT_UPDATED);
         }
+        const updatedTransaction = await this.getById(id);
+        this.notificationGateway.sendNotification(
+            transaction.userId,
+            TRANSACTION_UPDATED,
+            EVENT_TRANSACTION_UPDATED,
+            updatedTransaction,
+        );
         await this.redisService.client.del('transactions');
-        return await this.getById(id);
+        return updatedTransaction;
     }
 
     async deleteById(id: string): Promise<boolean> {
+        const transaction = await this.getById(id);
+        if (!transaction) {
+            throw new NotFoundException(TRANSACTION_NOT_FOUND);
+        }
         const affectedCount = await this.transactionModel.destroy({
             where: {
                 id,
@@ -81,6 +116,12 @@ export class TransactionsService {
         if (!affectedCount) {
             throw new BadRequestException(TRANSACTION_NOT_DELETED);
         }
+        this.notificationGateway.sendNotification(
+            transaction.userId,
+            TRANSACTION_DELETED,
+            EVENT_TRANSACTION_DELETED,
+            transaction,
+        );
         await this.redisService.client.del('transactions');
         return true;
     }
